VCV-128: restructure files to account for refactoring done in 150
